Prefix group with "gradle.plugin"
* Gradle used to automatically prefix the group id with "gradle.plugin"
but in the 0.13.0 release this was no longer happening, possibility
due to upgrading to Gradle 7.
* This group change caused an issue where the Gradle Plugin Portal
thought this plugin was being moved to a different group which would
affect the classpath and requires approval. However this is a breaking
change we don't want so we have to now pre-fix it ourselves.
* See the following issue for more context on why this is required now.
   - https://github.com/gradle/plugin-portal-requests/issues/103#issuecomment-859235098